+++ R/methods.R
@@ -47,7 +47,7 @@ sigma.feis <- function(object,...){
 }
 
 
-#' @title Calculate Variance-Covariance Matrix for a feis object
+#' @title Calculate Variance-Covariance Matrix for feis models
 #'
 #' @description
 #' Returns the variance-covariance matrix of the main parameters of an object of class "\code{feis}".
@@ -126,10 +126,57 @@ fitted.feis <- function(object,...){
 }
 
 
-#' @rdname feis
+
+
+#' @title Predict method for feis models
+#'
+#' @description
+#' Predicted values based on linear model object.
+#'
+#' @details
+#' \code{predict.lm} produces predicted values, obtained by evaluating the regression function
+#' in the frame newdata (which defaults to \code{model.matrix(object)}). If the logical \code{se.fit} is
+#' \code{TRUE}, standard errors of the predictions are calculated. If the \code{vcov} is not provided,
+#' the \code{object$vcov} is used, thus allowing for robust variance-covariance matrices.
+#' Setting intervals specifies computation of confidence or prediction (tolerance) intervals
+#' at the specified \code{level}.
+#'
+#' \bold{Note}: Currently, predictions are based on the transformed (detrended) data.
+#'
+#'
+#' @seealso \code{\link[stats]{predict.lm}, \link[stats]{predict}}
+#'
+#' @param object an object of class "\code{feis}", fitted model.
+#' @param newdata an optional data frame in which to look for variables with which to predict.
+#' If omitted, the fitted values are used.
+#' @param se.fit a switch indicating if standard errors are required.
+#' @param vcov optional variance-covariance matrix for std.err. calculation.
+#' @param interval type of interval calculation.
+#' @param level tolerance/confidence level.
+#' @param pred.var the variance for future observations to be assumed for prediction intervals.
+#' By default, equals the residual variance
+#' @param ...	further arguments.
+#'
+#' @return A vector of predictions or a matrix of predictions and bounds with column names
+#' \code{fit}, \code{lwr}, and \code{upr} if \code{interval} is set.
+#'
+#' @examples
+#' feis.mod <- feis(lnw ~ age | exp,
+#' data = mwp, id = "id", robust = TRUE)
+#'
+#' new <- data.frame(age = seq(-10, 10, 1))
+#' feis.pred <- predict(feis.mod, newdata = new,
+#'                      se.fit = TRUE, interval = "confidence")
+#' matplot(new$age, feis.pred$fit, lty = c(1,2,2),
+#'         type = "l", ylab = "predicted y")
+#'
+#'
 #' @export
-predict.feis <- function(object, newdata = NULL, ...){
+predict.feis <- function(object, newdata = NULL, se.fit = FALSE, vcov = NULL,
+                         interval = c("none", "confidence", "prediction"),
+                         level = 0.95, pred.var = sigma_sq, ...){
   f <- formula(object, rhs = 1, lhs = 0)
+  interval <- match.arg(interval)
   tt <- terms(f)
   if (is.null(newdata)){
     result <- fitted(object, ...)
@@ -151,6 +198,63 @@ predict.feis <- function(object, newdata = NULL, ...){
     result <- as.numeric(crossprod(beta, t(X)))
     names(result) <- rownames(X)
   }
+  if(se.fit == TRUE || interval != "none"){
+    vcov_arg <- object$vcov_arg
+    df <- df.residual(object)
+    sigma_sq <- sigma(object)^2
+    if(is.null(newdata)){
+      X <- model.matrix(object)[, all.vars(tt), drop = FALSE]
+    }
+    if(is.null(vcov)){
+      if(vcov_arg == "Cluster robust standard errors"){
+        vcov <- object$vcov[all.vars(tt), all.vars(tt), drop = FALSE]
+      } else {
+        vcov <- vcov(object, scale = TRUE)[all.vars(tt), all.vars(tt), drop = FALSE]
+      }
+    }
+    ip <- diag(X %*% tcrossprod(vcov, X))
+    se <- sqrt(ip)
+    names(se) <- names(result)
+    if(interval != "none"){
+      tfrac <- stats::qt((1 - level)/2, df)
+      hwid <- tfrac * switch(interval, confidence = se,
+                             prediction = sqrt(ip + pred.var))
+      result <- cbind(result, result + hwid %o% c(1, -1))
+      colnames(result) <- c("fit", "lwr", "upr")
+    }
+    result <- list(fit = result, se.fit = se, df = df, residual.scale = sqrt(sigma_sq))
+  }
+  # # Old version
+  # if(se.fit == TRUE || interval != "none"){
+  #   n <- length(object$residuals)
+  #   p <- length(object$assign)
+  #   p1 <- seq_len(p)
+  #   df <- df.residual(object)
+  #   vcov_arg <- object$vcov_arg
+  #   sigma_sq <- if(is.null(scale)){
+  #     sigma(object)^2
+  #   }else{scale^2}
+  #   if(is.null(vcov) && vcov_arg == "Normal standard errors"){
+  #     QR <- qr(model.matrix(object))
+  #     XRinv <- if (missing(newdata)){
+  #       qr.Q(QR)[, p1, drop = FALSE]
+  #     }else{
+  #       X %*% qr.solve(qr.R(QR))[p1, p1]
+  #     }
+  #     ip <- drop(XRinv^2 %*% rep(sigma_sq, p))
+  #     se <- sqrt(ip)
+  #   } else{
+  #     stop("Currently only for conventional standard errors")
+  #   }
+  #   if(interval != "none"){
+  #     tfrac <- qt((1 - level)/2, df)
+  #     hwid <- tfrac * switch(interval, confidence = sqrt(ip),
+  #                            prediction = sqrt(ip + pred.var))
+  #     result <- cbind(result, result + hwid %o% c(1, -1))
+  #     colnames(result) <- c("fit", "lwr", "upr")
+  #   }
+  #   result <- list(fit = result, se.fit = se, df = df, residual.scale = sqrt(sigma_sq))
+  # }
   result
 }
 
@@ -188,6 +292,7 @@ model.matrix.feis <- function(object, ...){
   data <- data[, which(colnames(data) != resp), drop = FALSE]
   return(data)
 }
+